+++ src/serializers/simple.rs
@@ -1,5 +1,6 @@
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
+
 use serde::Serialize;
 
 use super::any::{fallback_serialize, fallback_to_python, IsType, ObType};
