+++ src/serializers/tuple.rs
@@ -8,8 +8,9 @@ use crate::build_tools::SchemaDict;
 
 use super::any::{fallback_serialize, fallback_to_python, AnySerializer};
 use super::include_exclude::SchemaIncEx;
-use super::shared::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, SerMode, TypeSerializer};
-use super::PydanticSerializer;
+use super::shared::{
+    py_err_se_err, BuildSerializer, CombinedSerializer, Extra, PydanticSerializer, SerMode, TypeSerializer,
+};
 
 pub struct TupleBuilder;
