+++ src/serializers/url.rs
@@ -1,10 +1,12 @@
+use std::borrow::Cow;
+
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
-use std::borrow::Cow;
+
+use crate::url::{PyMultiHostUrl, PyUrl};
 
 use super::any::{fallback_serialize, fallback_to_python_json, json_key};
 use super::shared::{BuildSerializer, CombinedSerializer, Extra, SerMode, TypeSerializer};
-use crate::url::{PyMultiHostUrl, PyUrl};
 
 macro_rules! build_serializer {
     ($struct_name:ident, $expected_type:literal, $extract:ty) => {
